Make container settable in object path.
Make it possible to change the container a file will be uploaded too
using it path.

Change the path schema slightly, now if you use a path with a colon in
it, it will use the text before the colon to select the container e.g.

"container:some/path/to/file.jpg"

will place the file in the container named "container".

Not adding a container section will result in the file being uploaded
to the default container.

To match there has also been a change in the way the settings work.

There is now a "CONTAINERS" section which is a dictionary of legal
container names and whether they should be public or not.

There is also a "DEFAULT_CONTAINER" option for setting the name of the
default container (must be in "CONTAINERS").

Additionally if the container does not exist then we will create it the
first time it is used.

Finally I have added a few new test to use the new features, however
they are not complete and this feature need a little more polishing.